# docs/swagger.json
{
    "swagger": "2.0",
    "info": {
        "description": "E-commerce application (Product: Shoes)",
        "title": "MyShoo API",
        "contact": {},
        "version": "1.0"
    },
    "paths": {
...
        "/seller/login": {
            "get": {
                "security": [
                    {
                        "BearerTokenAuth": []
                    }
                ],
                "description": "Seller Login Page",
                "consumes": [
                    "application/json"
                ],
                "produces": [
                    "application/json"
                ],
                "tags": [
                    "Seller/Session"
                ],
                "summary": "Seller Login Page",
                "responses": {
                    "200": {
                        "description": "OK",
                        "schema": {
                            "type": "string"
                        }
                    }
                }
            },
            "post": {
                "security": [
                    {
                        "BearerTokenAuth": []
                    }
                ],
                "description": "Seller Sign In Handler",
                "consumes": [
                    "application/json"
                ],
                "produces": [
                    "application/json"
                ],
                "tags": [
                    "Seller/Session"
                ],
                "summary": "Seller Sign In Handler",
                "parameters": [
                    {
                        "description": "Seller Sign In Request",
                        "name": "req",
                        "in": "body",
                        "required": true,
                        "schema": {
                            "$ref": "#/definitions/req.SellerSignInReq"
                        }
                    }
                ],
                "responses": {
                    "200": {
                        "description": "OK",
                        "schema": {
                            "$ref": "#/definitions/response.SMT"
                        }
                    },
                    "400": {
                        "description": "Bad Request",
                        "schema": {
                            "$ref": "#/definitions/response.SME"
                        }
                    }
                }
            }
        },
...
        "/seller/signup": {
            "post": {
                "security": [
                    {
                        "BearerTokenAuth": []
                    }
                ],
                "description": "Seller Sign Up Handler",
                "consumes": [
                    "application/json"
                ],
                "produces": [
                    "application/json"
                ],
                "tags": [
                    "Seller/Session"
                ],
                "summary": "Seller Sign Up Handler",
                "parameters": [
                    {
                        "description": "Seller Sign Up Request",
                        "name": "req",
                        "in": "body",
                        "required": true,
                        "schema": {
                            "$ref": "#/definitions/req.SellerSignUpReq"
                        }
                    }
                ],
                "responses": {
                    "200": {
                        "description": "OK",
                        "schema": {
                            "type": "string"
                        }
                    },
                    "400": {
                        "description": "Bad Request",
                        "schema": {
                            "type": "string"
                        }
                    }
                }
            }
        },
...
    },
    "definitions": {
...
        "req.SellerSignInReq": {
            "type": "object",
            "required": [
                "Login_email",
                "Login_password"
            ],
            "properties": {
                "Login_email": {
                    "type": "string"
                },
                "Login_password": {
                    "type": "string",
                    "minLength": 3
                }
            }
        },
        "req.SellerSignUpReq": {
            "type": "object",
            "required": [
                "SignUp_email",
                "SignUp_firstName",
                "SignUp_phone"
            ],
            "properties": {
                "SignUp_email": {
                    "type": "string"
                },
                "SignUp_firstName": {
                    "type": "string",
                    "minLength": 3
                },
                "SignUp_lastName": {
                    "type": "string"
                },
                "SignUp_password": {
                    "description": "1st password shall only be considered. Its frontend to check both typed passwords are same. No security implicatoions",
                    "type": "string",
                    "minLength": 3
                },
                "SignUp_phone": {
                    "type": "string"
                }
            }
        },
...
        "response.SME": {
            "type": "object",
            "properties": {
                "error": {
                    "type": "string"
                },
                "message": {
                    "type": "string"
                },
                "status": {
                    "type": "string"
                }
            }
        },
...
        "response.SMT": {
            "type": "object",
            "properties": {
                "message": {
                    "type": "string"
                },
                "status": {
                    "type": "string"
                },
                "token": {
                    "type": "string"
                }
            }
        },
...
    },
    "securityDefinitions": {
        "BearerTokenAuth": {
            "type": "apiKey",
            "name": "Authorization",
            "in": "header"
        }
    }
}